requirements.txt
flask==3.0.2
flask-socketio==5.3.6
flask-cors==4.0.0
google-generativeai==0.7.1
SpeechRecognition==3.10.1
pydub==0.25.1
python-engineio==4.9.0
python-socketio==5.11.0
eventlet==0.35.2 